Fold edit-handling functionality into data-container
The intention is to wrap components using withCopyEdit and to use
a 'mode' prop to control view/edit/copy/paste mode.

Next steps to unpack form field data components and update in eaech
category container.
